[migclic] Add send_fiche_to_instance()
Add functionality to send the newly created fiches from google contacts
to the clicrdv instance.

Add --Create argument to confirm real creation of new fiches. It
defaults to False so the fiches are only printed out but not created.

The 'real' send code is commented out until we get access to the sandbox
instance for write tests.

Signed-off-by: Louis Bouchard <louis@ubuntu.com>